Fetching User for each blog and show the name. (with problem)
Problem: For each blog, the data of the user is fetched. That means, that a user is fetched for a blog post, even when the use was already fetched (user has more than one blog post). Concretely, for this application, each user is fetched 10 times (100 blogs with 10 users, each 10 blogs) which means, we fetch data from the server 90 times more than neccessary.